feat: its alive!

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -21,7 +21,7 @@ public class FilmController {
     }
 
     @GetMapping("/{id}")
-    public Film findFilm(@PathVariable("id") int id) {
+    public Film findFilm(@PathVariable("id") long id) {
         return filmService.findFilm(id);
     }
 
@@ -46,17 +46,17 @@ public class FilmController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteFilm(@PathVariable("id") int id) {
+    public void deleteFilm(@PathVariable("id") long id) {
         filmService.deleteFilm(id);
     }
 
     @PutMapping("/{id}/like/{userId}")
-    public void addLike(@PathVariable("id") int id, @PathVariable("userId") int userId) {
+    public void addLike(@PathVariable("id") long id, @PathVariable("userId") long userId) {
         filmService.addLike(id, userId);
     }
 
     @DeleteMapping("/{id}/like/{userId}")
-    public void deleteLike(@PathVariable("id") int id, @PathVariable("userId") int userId) {
+    public void deleteLike(@PathVariable("id") long id, @PathVariable("userId") long userId) {
         filmService.deleteLike(id, userId);
     }
